+++ meme_generator_memes/src/memes.rs
@@ -7,7 +7,7 @@ mod alike;
 mod alipay;
 mod always;
 mod always_like;
-mod anan;
+mod anan_say;
 mod anti_kidnap;
 mod anya_suki;
 mod applaud;
@@ -65,6 +65,7 @@ mod dont_touch;
 mod doraemon_say;
 mod douyin;
 mod eat;
+mod elysia_holdsign;
 mod erised_mirror;
 mod fade_away;
 mod fanatic;
@@ -280,4 +281,3 @@ mod you_dont_get;
 mod you_should_call;
 mod your_new_years_eve;
 mod youtube;
-mod elysia_holdsign;
